Add TablesEvents management functionality
- Implemented `TablesController` for CRUD operations on `TablesEvents`.
- Updated `DataContext` to include `DbSet<TablesEvents>`.
- Created migration files to add `TablesEvents` table and update `Invitations`.
- Registered new services in `Program.cs` for dependency injection.
- Enhanced `InvitationRepository` to manage relationships with `TablesEvents`.
- Added `TablesEventsRepository` for repository pattern implementation.
- Created interfaces `ITablesEventsRepository` and `ITablesEventsUnitOfWork`.
- Updated UI components in `EventsDetails.razor` for table selection and assignment.
- Developed `TablesIndex.razor` for managing tables and their invitations.
- Added DTOs in `TablesDto.cs` for table creation and editing.
- Updated `Event` and `Invitation` entities to include navigation properties for `TablesEvents`.
- Defined `TablesEvents` entity with necessary properties and relationships.

+++ SIC/SIC.Backend/Repositories/Implemetations/UserRepository.cs
@@ -35,11 +35,11 @@ namespace SIC.Backend.Repositories.Implemetations
             }
 
             double count = await queryable.CountAsync();
-            double totalPages = Math.Ceiling(count / pagination.PageSize);
+            int totalPages = (int)Math.Ceiling(count / pagination.PageSize);
             return new ActionResponse<int>
             {
                 Success = true,
-                Result = (int)totalPages
+                Result = totalPages
             };
         }
